benchmarks: fix issues in benchmarks (ast_evaluator.py)
Changes:
- ast_evaluator.py: `isinstance` does not accept PEP-604 unions (`int | float`) as the second arg...

+++ benchmarks/bfcl/evaluators/ast_evaluator.py
@@ -225,7 +225,7 @@ class ASTEvaluator:
         result = " ".join(result.split())
         return result.lower()
 
-    def _try_parse_number(self, value: object) -> Optional[int | float]:
+    def _try_parse_number(self, value: object) -> Optional[float | int]:
         """Try to parse a value as a number."""
         if isinstance(value, (int, float)) and not isinstance(value, bool):
             return value
